fix: transaction id url & member clear all dues action url

+++ app/books/views/delete_book.py
@@ -7,7 +7,11 @@ from app.books.models import Book
 def delete_book(id: int):
     book: Book = db.get_or_404(Book, id)
     book_id: int = book.id
-    db.session.delete(book)
-    db.session.commit()
-    flash(f"Book ID - {book_id} deleted successfully.", "success")
+
+    if not book.book_safe_to_delete:
+        flash(f"Book ID - {book_id} cannot be deleted as stocks are rented.", "danger")
+    else:
+        db.session.delete(book)
+        db.session.commit()
+        flash(f"Book ID - {book_id} deleted successfully.", "success")
     return redirect(url_for('books.books'))
