+++ app/books/views/delete_stock.py
@@ -1,7 +1,7 @@
 from sqlalchemy import and_
 from flask import abort, redirect, flash, url_for
 from app.db import db
-from app.books.models import BookStock
+from app.books.models import BookStock, StockStatus
 from app.books.base import books_bp
 
 @books_bp.route('/<int:bid>/<int:sid>/delete', methods=['POST'])
@@ -21,7 +21,11 @@ def delete_stock(bid: int, sid :int):
 
     stock_id: int = book_stock.id
     book_id: int = book_stock.book_id
-    db.session.delete(book_stock)
-    db.session.commit()
-    flash(f"Stock ID - {stock_id} deleted successfully.", "success")
+
+    if book_stock.status == StockStatus.TAKEN:
+        flash(f"Stock ID - {stock_id} is currently rented.", "danger")
+    else:    
+        db.session.delete(book_stock)
+        db.session.commit()
+        flash(f"Stock ID - {stock_id} deleted successfully.", "success")
     return redirect(url_for('books.book_view', id=book_id))
