+++ app/members/views/delete_member.py
@@ -7,7 +7,11 @@ from app.members.base import members_bp
 def delete_member(id: int):
     member: Member = db.get_or_404(Member, id)
     member_id = member.id
-    db.session.delete(member)
-    db.session.commit()
-    flash(f'Member - {member_id} deleted successfully.', 'success')
+
+    if member.no_of_books_currently_rented:
+        flash(f'Member - {member_id} have to return book before deletion.', 'danger')
+    else:
+        db.session.delete(member)
+        db.session.commit()
+        flash(f'Member - {member_id} deleted successfully.', 'success')
     return redirect(url_for('members.members'))
